--- README.md
@@ -1,36 +1,0 @@
-### One File Home-task view:
-`lesson_<number>/<surname>_<home-task_number>.py`\
-lesson_04/maksymchuk_04.py
-
-### Several Files Home-task view:
-`lesson_<number>/<surname>/*`\
-lesson_04/maksymchuk/*
-
-## Workflow how to work with Git.
-Before you start to do your task (all examples for lesson_04):
-1. Be sure you are in repository folder and pull all recent changes form `hometasks` branch with command:
-   `git pull origin hometasks`
-
-2. Create and switch to your new branch.\
-   Branch name must be like: `<surname>/lesson_<number>` :\
-    `git checkout -b maksymchuk/lesson_04`
-
-3. Do your home task carefully.
-4. Add all newly created files with `git add` command.
-5. Make commit with `git commit -m "lesson_04 done"`
-6. Push your changes `git push origin maksymchuk/lesson_04`
-7. Go to `https://github.com/MaximTwister/python-22-1/` \
-   You can see the invitation to make a PR with your new branch:\
-   https://i.ibb.co/R6JSd52/make-pr.png
-
-8. Press "Compare & pull request" button
-9. Make sure you are going to merge `maksymchuk/lesson_04` branch to `hometasks` branch:\
-   https://i.ibb.co/LQbnwRt/choose-branch.png
-
-10. Press "Create pull request" button
-11. Wait while your PR will be approved:\
-    https://i.ibb.co/4fMCqTq/wait-for-pr.png
-
-12. After your PR will be approved - press green button "Merge pull request".
-13. Delete your branch `maksymchuk/lesson_04` from GitHub repository.
-14. Now you are good.
