--- lesson_04/Rysich_4.py
@@ -1,1 +1,0 @@
-print("Test Yaroslav")
