--- lesson_04/kaliuzhna_04.py
@@ -1,3 +1,0 @@
-print("Hello, GIT!")
-
-
