--- lesson_04/mulik_04.py
@@ -1,1 +1,0 @@
-print("hello world")
