--- lesson_04/serhiienko_04.py
@@ -1,1 +1,0 @@
-print("hello world!!!")
