--- lesson_04/task.py
@@ -1,1 +1,0 @@
-# Create your very first commit to branch `hometasks` with any python file.
