--- lesson_06/kaliuzhna/constants.py
@@ -1,5 +1,0 @@
-EMPTY_STRING = ""
-FILENAME_LENGTH = 5  # count symbols in file name  > 0, int
-FILES_AMOUNT = 20  # number of similar files
-EXTENSIONS = ["txt", "mp3", "pdf", "docx", "py", "tmp"]
-WORKING_DIR = "./tania_test"
