--- lesson_07/generator_maintanance/README.md
@@ -1,20 +1,0 @@
-Menu structure:
-
-# 1. Add new generator
-#   1.1 Name -> Generator in room 230
-#   1.2 Description
-#   1.3 Model -> GS-305
-#   1.4 Change oil period (in hours)
-#
-# 2. List generators
-#   shows us all our generators -> int
-#   2.1 Status
-#   2.2 History
-#
-# 3. Start
-#   3.1 *Shows us generator list -> int
-#
-# 4. Stop
-#   4.1 Shows us generator list in status "Working" -> int
-#
-# 5. Exit
