--- lesson_07/generator_maintanance/constants.py
@@ -1,2 +1,0 @@
-GEN_KEYS = ["name", "description", "model", "change oil (motohours)"]
-GEN_DB_FILE = "./gen_db"
