--- lesson_07/generator_maintanance/gen_db
@@ -1,1 +1,0 @@
-[{"name": "Test", "description": "Test descr", "model": "TT-11", "change oil (motohours)": "11", "id": 1}]
