--- lesson_07/task.py
@@ -1,3 +1,0 @@
-# 1. Read code many times
-# 2. https://pypi.org/project/prettytable/
-# 3. generator's `id` key should be on the first place
